Added: ExceptionHandlerController, Exceptions

+++ src/main/java/com/example/aviasimbir/service/FlightService.java
@@ -3,18 +3,19 @@ package com.example.aviasimbir.service;
 import com.example.aviasimbir.entity.Flight;
 import com.example.aviasimbir.entity.Logger;
 import com.example.aviasimbir.entity.Plane;
+import com.example.aviasimbir.exceptions.FlightTimeException;
+import com.example.aviasimbir.exceptions.NoSuchIdException;
+import com.example.aviasimbir.exceptions.WrongArgumentException;
 import com.example.aviasimbir.repo.FlightRepository;
 import com.example.aviasimbir.repo.LoggerRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.StringUtils;
 
 import java.time.Instant;
 import java.time.ZonedDateTime;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -34,10 +35,16 @@ public class FlightService {
      * @param id идентификатор рейса
      * @return рейс
      */
-    public Optional<Flight> findFlight(Long id) {
-        log.info("IN getFlight - Flight: {} successfully found", id);
-        return flightRepository.findById(id);
+    public Flight getFlight(Long id) throws NoSuchIdException {
+        Optional<Flight> flightOptional = flightRepository.findById(id);
+        if (flightOptional.isPresent()) {
+            log.info("IN getFlight - Flight: {} successfully found", id);
+            return flightOptional.get();
+        } else {
+            throw new NoSuchIdException("Flight with id " + id + " was not found");
+        }
     }
+
 
     /**
      * Получение всех рейсов
@@ -59,11 +66,23 @@ public class FlightService {
      */
     @Transactional
     public Flight createFlight(Plane plane, String departure, String destination,
-                               ZonedDateTime departureTime, ZonedDateTime arrivalTime) {
-        Flight flight = new Flight(plane, departure, destination, departureTime, arrivalTime);
-        flightRepository.save(flight);
-        log.info("IN createFlight - Flight: {} successfully created", flight.getId());
-        return flight;
+                               ZonedDateTime departureTime, ZonedDateTime arrivalTime) throws WrongArgumentException, FlightTimeException {
+        if (plane == null || departure == null || departure.trim().isEmpty() || departureTime == null || arrivalTime == null || destination == null || destination.trim().isEmpty()) {
+            throw new WrongArgumentException("All fields must be filled");
+        } else {
+            if (departureTime.isAfter(arrivalTime)) {
+                throw new FlightTimeException("Departure time cannot be after arrival time");
+            } else {
+                if (departureTime.isBefore(ZonedDateTime.now())) {
+                    throw new FlightTimeException("Departure time cannot be before current time");
+                } else {
+                    Flight flight = new Flight(plane, departure, destination, departureTime, arrivalTime);
+                    flightRepository.save(flight);
+                    log.info("IN createFlight - Flight: {} successfully created", flight.getId());
+                    return flight;
+                }
+            }
+        }
     }
 
     /**
@@ -78,35 +97,27 @@ public class FlightService {
      * @return рейс
      */
     @Transactional
-    public Optional<Flight> updateFlight(Long id, ZonedDateTime departureTime,
-                                         ZonedDateTime arrivalTime, Plane plane,
-                                         String departure, String destination) {
+    public Flight updateFlight(Long id, ZonedDateTime departureTime,
+                               ZonedDateTime arrivalTime, Plane plane,
+                               String departure, String destination) throws WrongArgumentException, NoSuchIdException {
         Optional<Flight> optionalFlight = flightRepository.findById(id);
-        if (optionalFlight.isEmpty()) {
-            return Optional.empty();
+        if (optionalFlight.isPresent()) {
+            if (plane == null || departure == null || departure.trim().isEmpty() || departureTime == null || arrivalTime == null || destination == null || destination.trim().isEmpty()) {
+                throw new WrongArgumentException("All fields must be filled");
+            } else {
+                Flight flight = optionalFlight.get();
+                flight.setPlane(plane);
+                flight.setDepartureTime(departureTime);
+                flight.setArrivalTime(arrivalTime);
+                flight.setDeparture(departure);
+                flight.setDestination(destination);
+                flightRepository.save(flight);
+                log.info("IN updateFlight - Flight: {} successfully updated", id);
+                return flight;
+            }
+        } else {
+            throw new NoSuchIdException("Flight with id " + id + " was not found");
         }
-        Flight flight = optionalFlight.get();
-        if (Objects.nonNull(departureTime)) {
-            flight.setDepartureTime(departureTime);
-        }
-        if (Objects.nonNull(arrivalTime)) {
-            flight.setArrivalTime(arrivalTime);
-        }
-        if (Objects.nonNull(plane)) {
-            flight.setPlane(plane);
-        }
-        if (StringUtils.hasText(departure)) {
-            flight.setDeparture(departure);
-        }
-        if (StringUtils.hasText(destination)) {
-            flight.setDestination(destination);
-        }
-        if (!flightRepository.existsById(flight.getId())) {
-            return Optional.empty();
-        }
-        flightRepository.save(flight);
-        log.info("IN updateFlight - Flight: {} successfully updated", id);
-        return optionalFlight;
     }
 
     /**
@@ -115,10 +126,17 @@ public class FlightService {
      * @param id идентификатор рейса
      */
     @Transactional
-    public void deleteFlight(Long id) {
-        flightRepository.deleteById(id);
-        log.info("IN deleteFlight - Flight: {} successfully deleted", id);
-        loggerRepository.save(new Logger("Flight " + id + " was deleted", Instant.now()));
+    public void deleteFlight(Long id) throws NoSuchIdException {
+        if (flightRepository.findById(id).isPresent()) {
+
+
+            flightRepository.deleteById(id);
+            log.info("IN deleteFlight - Flight: {} successfully deleted", id);
+            loggerRepository.save(new Logger("Flight " + id + " was deleted", Instant.now()));
+        } else {
+            throw new NoSuchIdException("Flight with id " + id + " was not found");
+        }
+
     }
 
     /**
